requirements.txt
crcmod==1.7
python-memcached==1.58
RPi.GPIO==0.6.2
six==1.10.0
spidev==3.2
